Add per-plugin order and error callbacks; use Decimal for TradeSignal quantity
- TradeSignal.quantity changed from int to Decimal; update all usages
  across plugins, order_reconciler, plugin_executive, and tests
- Route IB orderStatus callbacks per-plugin: new on_order_fill and
  on_order_status hooks on PluginBase; executive wires portfolio callback
  and dispatches to owning plugin(s) via _order_id_to_plugins map
- Route IB error callbacks per-plugin: new on_ib_error hook; executive
  routes by order ID first, then by stream symbol; filters req_id==-1
  and informational codes (2104/2106/2158/2119/10167)
- register_order() helper on PluginBase for directly-placed orders
- paper_tests: replace time.sleep polling with threading.Event wakeup;
  add terminal-state early exit in _wait_fill_cancel_other
- Add tests/test_callbacks.py with 55 tests covering all three changes
- Update PLUGIN_MANUAL.md: Decimal quantity docs, new Section 13 (Order
  Callbacks and Error Routing), threading notes, renumber sections 13-20

+++ tests/test_order_reconciler.py
@@ -4,6 +4,7 @@ Tests for order_reconciler.py - Order reconciliation and netting
 
 import pytest
 from datetime import datetime
+from decimal import Decimal
 from unittest.mock import Mock, MagicMock
 from ibapi.contract import Contract
 from ibapi.order import Order
@@ -28,12 +29,12 @@ def create_contract(symbol: str) -> Contract:
     return contract
 
 
-def create_signal(symbol: str, action: str, quantity: int, confidence: float = 0.8) -> TradeSignal:
+def create_signal(symbol: str, action: str, quantity: Decimal, confidence: float = 0.8) -> TradeSignal:
     """Helper to create a test trade signal"""
     return TradeSignal(
         symbol=symbol,
         action=action,
-        quantity=quantity,
+        quantity=Decimal(quantity),
         confidence=confidence,
         reason="test signal",
     )
